Fixes: when auto metadata extraction feature is disabled, the format metadata field should still be auto-prefilled
