+++ CHANGES.md
@@ -5,6 +5,7 @@
 - #3581 Make storage API upload limit configurable
 - #3582 Add cache support to SQLConsole
 - Upgrade dompurify to v3.2.4
+- Fixes: when auto metadata extraction feature is disabled, the format metadata field should still be auto-prefilled
 
 ## v5.0.0
